completed the todolist api

+++ src/modules/tasks/tasks.service.ts
@@ -1,12 +1,14 @@
 import { CreateTask } from './task.interface'
 import TaskModel from './Task.model'
+import userModel from '../user/user.model'
 
-const createTask = async (createTaskData: CreateTask) => {
-  return await TaskModel.create(createTaskData)
+const createTask = async (userData:any,createTaskData: CreateTask) => {
+  return await TaskModel.create({...createTaskData,userId:userData._id})
 }
 
-const getAllTasks = async()=>{
-  return await TaskModel.find()
+const getAllTasks = async(userId:string)=>{
+return await TaskModel.find({userId})
+  
 }
 
 const getTaskById = async(taskId:string)=>{
@@ -20,8 +22,15 @@ const updateTask=async(taskId:string,taskData:Partial<CreateTask>)=>{
 }
 
 const deleteTask = async(taskId:string)=>{
+  const task = await TaskModel.findById(taskId)
+  if(!task){
+    throw new Error("Task not found")
+  }
+
   return await TaskModel.deleteOne({taskId})
 }
+
+
 
 export const taskService = {
   createTask,
